requirements.txt
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
numpy>=1.21.0
opencv-python>=4.5.0
Pillow>=8.3.0
scikit-learn>=1.0.0
scikit-image>=0.19.0
matplotlib>=3.5.0
seaborn>=0.11.0
tqdm>=4.62.0
pyyaml>=6.0
tensorboard>=2.8.0
wandb>=0.12.0
albumentations>=1.2.0
timm>=0.6.0
efficientnet-pytorch>=0.7.0
pandas>=1.3.0
scipy>=1.7.0
h5py>=3.6.0
imageio>=2.16.0
pytest>=6.2.0
pytest-cov>=3.0.0
black>=22.0.0
flake8>=4.0.0
isort>=5.10.0
pre-commit>=2.17.0
tensorflow>=2.8.0